cleanup: Move debug logging code to separate file.
